users authorisation req - only admin. delete admin can all.

+++ app_api/controllers/comments.js
@@ -8,6 +8,19 @@ var JSONcallback = function(res, status, msg) {
   res.json(msg);
 };
 
+var decodeAndVerify = function(req, res){
+    var authorization = req.headers.authorization,
+            decoded;
+        authorization = authorization.split(' ')[1];
+        try {
+            decoded = jwt.verify(authorization, process.env.JWT_PASS);
+        } catch (e) {
+            return false;
+        }
+        
+        return decoded;
+};
+
 var createComment = function(req,res,user,datatime){
     /*
         The .save() is an instance method of the model, while the 
@@ -19,10 +32,11 @@ var createComment = function(req,res,user,datatime){
         name: req.body.name,
         comment: req.body.comment,
         pic: req.body.pic,
-        date: datatime
+        date: datatime,
+        admin: user.admin // did admin created the comment
     }, function(error, comment){
         if(error){
-            JSONcallback(res,404,{
+            JSONcallback(res,500,{
                 msg: 'Error while storing the comment.'+
                 ' Comment not stored.'
             });
@@ -56,12 +70,12 @@ var deleteCommentFromUser = function(res,comment){
                     'Comment is deleted but the user\'s ref has stayed. '+
                     'Inconsistant state of the user.'+user.mail
                     });
-                }
+                } else {
                 console.log(user," is removed to the list of your comments");
                 JSONcallback(res, 200, {
                     removed: comment
                     
-                });
+                });}
             });
 };
 
@@ -70,7 +84,7 @@ module.exports.getAll = function(req, res) {
     .exec(function(err, comment){
         if (err) {
             console.log(err);
-            JSONcallback(err, 400, comment);
+            JSONcallback(err, 500, comment);
         }else{
             JSONcallback(res, 200, comment);
         }
@@ -86,7 +100,7 @@ module.exports.createNew = function(req, res) {
         and comments needed data.*/
     if( !req.body.username || !req.body.name 
             || !req.body.comment){
-        JSONcallback(res, 400, {
+        JSONcallback(res, 500, {
           msg: "All data req."
         });
         return;
@@ -96,13 +110,13 @@ module.exports.createNew = function(req, res) {
             mail: req.body.username
         }).exec(function(error,user){
             if(error){
-                JSONcallback(res, 404, {
+                JSONcallback(res, 500, {
                   msg: error
                 });
                 return;
             }
             if(!user){
-                JSONcallback(res, 401, {
+                JSONcallback(res, 500, {
                   msg: "No user with username:"+req.body.username
                   +". Comment can not be added."
                 });
@@ -118,7 +132,7 @@ module.exports.getCommentByName = function(req, res) {
         name: req.query.name
     }, function(error, data){
         if(error){
-            JSONcallback(res,400,error);
+            JSONcallback(res,500,error);
         } else {
             JSONcallback(res, 200, data);
         }
@@ -129,7 +143,7 @@ module.exports.getCommentById = function(req, res) {
     Comment.findById(
          req.params.idComment, function(error, data){
         if(error){
-            JSONcallback(res,400,error);
+            JSONcallback(res,500,error);
         } else {
             JSONcallback(res, 200, data);
         }
@@ -139,40 +153,53 @@ module.exports.getCommentById = function(req, res) {
 module.exports.deleteCommentById = function(req, res) {
     //get header auth token.
     if (req.headers && req.headers.authorization) {
-        var authorization = req.headers.authorization,
-            decoded;
-        authorization = authorization.split(' ')[1];
-        try {
-            decoded = jwt.verify(authorization, process.env.JWT_PASS);
-        } catch (e) {
-            return res.status(401).send('unauthorized');
+        var decoded = decodeAndVerify(req, res);
+        if(!decoded) return JSONcallback(res,401,{msg: 'Unauthorized'});
+        
+        //admin can do anything
+        if(decoded.admin){
+            Comment
+            .findOneAndRemove({
+                _id: req.params.idComment 
+            })
+            .exec(function (error, comment) {
+                if (error) {
+                    JSONcallback(res,500,{
+                      msg: 'Error while deleting the comment. Comment might be in incosistent state.',
+                      error: error
+                  });
+                  return;
+                }
+                
+                JSONcallback(res,200,comment);
+            });
+            return;
         }
+        
+        //check if user has a comment
+        Comment
+        .findOneAndRemove({ 
+            _creator: decoded._id,
+            _id: req.params.idComment 
+        })
+        //populate create to fatch the user
+        .populate('_creator')
+        .exec(function (error, comment) {
+        if (error) {
+          JSONcallback(res,500,{
+              msg: 'Error.',
+              error: error
+          });
+          return;
+        } 
+        if(!comment) {
+            JSONcallback(res,401,{
+              msg: 'No comment connected with logedin user.'
+          });
+        } else {
+            deleteCommentFromUser(res,comment);
+        }});
     }
-    //user info in decoded
-    //console.log(decoded);
-    
-    //check if user has a comment
-    Comment
-    .findOneAndRemove({ 
-        _creator: decoded._id,
-        _id: req.params.idComment 
-    })
-    //populate create to fatch the user
-    .populate('_creator')
-    .exec(function (error, comment) {
-    if (error) {
-      JSONcallback(res,400,{
-          msg: 'Error.',
-          error: error
-      });
-    } 
-    if(!comment) {
-        JSONcallback(res,401,{
-          msg: 'No comment connected with logedin user.'
-      });
-    } else {
-        deleteCommentFromUser(res,comment);
-    }});
 };
 
 //suppode to have req.params.idComment & req.body.comment
@@ -193,7 +220,7 @@ module.exports.editComment = function(req, res) {
          {new: true},
          function(error, data){
             if(error){
-                JSONcallback(res,400,error);
+                JSONcallback(res,500,error);
             } else {
                 if (error) return console.error(error);
                     JSONcallback(res, 200, data);
@@ -201,4 +228,8 @@ module.exports.editComment = function(req, res) {
             }
         });
 };
+
+//if non-admin becomes admin, tehre is no need to update its comments before becoming admin as they are irrelevent,
+//it was user and as such posted comments. Only new one are commencted to its position,
+//as admin.
